pyproject.toml
[project]
name = "ecoshop-streams"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "confluent-kafka>=2.11.0",
    "faker>=37.5.3",
    "psycopg2-binary>=2.9.10",
]
